able to change duty cycle
