+++ Src/platform_init.h
@@ -18,6 +18,7 @@
 #include <UART/uart.h>
 #include <print_scan.h>
 #include <PWM/pwm_timer.h>
+#include <Intercomm/mgt_handler.h>
 
 
 /**
@@ -34,9 +35,7 @@ void init_platform() {
 	led_init();
 	printer_init();
 	systick_init();
-	EXTI->IMR1 |= 1;
-	NVIC_SetPriority(EXTI0_IRQn, 15);
-	NVIC_EnableIRQ(EXTI0_IRQn);
+	init_mgtHandler();
 }
 
 #endif // REALOP1_PLATFORM_INIT_H
